refactor: state -> states
